Safari-611.1.10/LayoutTests/http/tests/navigation/anchor-frames-cross-origin-expected.txt



--------
Frame: 'child'
--------
This prevents a cross-origin information leak sometimes know as framesniffing.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.scrollingElement.scrollTop == 0 is true
PASS document.scrollingElement.scrollLeft == 0 is true
PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: 'grandchild'
--------
This is an anchor point named "anchor1".
